pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cortex-engine"
version = "0.1.0"
description = "Real-time inference engine for transformer-based neural decoders"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Peter" }]

dependencies = [
  # Core ML
  "torch>=2.5.0",
  "triton>=3.0.0",
  "einops>=0.8.0",
  "numpy>=1.26",
  "scipy>=1.13",

  # Config and CLI
  "hydra-core>=1.3.2",
  "omegaconf>=2.3",
  "typer>=0.12",

  # Data
  "pandas>=2.2",
  "pynwb>=2.8",
  "dandi>=0.62",
  "h5py>=3.11",

  # Validation and serialization
  "pydantic>=2.7",
  "pydantic-settings>=2.3",

  # Training infrastructure
  "wandb>=0.17",
  "tqdm>=4.66",

  # Serving
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "websockets>=12.0",
  "httpx>=0.27",

  # Observability
  "prometheus-client>=0.20",
  "opentelemetry-api>=1.25",
  "opentelemetry-sdk>=1.25",
  "opentelemetry-instrumentation-fastapi>=0.46b0",
  "structlog>=24.1",

  # Utilities
  "rich>=13.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "pytest-benchmark>=4.0",
  "pytest-cov>=5.0",
  "hypothesis>=6.103",
  "mypy>=1.10",
  "ruff>=0.5",
  "black>=24.4",
  "isort>=5.13",
  "pre-commit>=3.7",
  "types-pyyaml",
  "types-requests",
]

baselines = [
  "scikit-learn>=1.5",
]

notebooks = [
  "jupyter>=1.0",
  "ipykernel>=6.29",
  "matplotlib>=3.9",
  "seaborn>=0.13",
]

load-test = [
  "locust>=2.29",
]

[project.scripts]
cortex-train = "cortex.training.cli:main"
cortex-serve = "cortex.serve.cli:main"
cortex-bench = "cortex.benchmarks.cli:main"

[tool.setuptools.packages.find]
include = ["cortex*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "B",   # flake8-bugbear
  "UP",  # pyupgrade
  "SIM", # flake8-simplify
  "RUF", # ruff-specific
]
ignore = [
  "E501",  # line length handled by black
  "B008",  # function calls in defaults (FastAPI Depends pattern)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]
"notebooks/*" = ["E402"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["cortex"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
no_implicit_optional = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
  "triton.*",
  "wandb.*",
  "pynwb.*",
  "dandi.*",
  "umap.*",
  "sklearn.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
  "slow: deselect with '-m \"not slow\"'",
  "gpu: requires CUDA",
  "integration: integration tests requiring full stack",
]
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
]

[tool.coverage.run]
source = ["cortex"]
omit = ["cortex/benchmarks/*", "*/tests/*"]
